viewmodelファイルに分けるようにしてて、アプリちゃんと起動時に表示されるようにした
