+++ TitechAppLite/ContentView.swift
@@ -10,37 +10,7 @@ import SwiftUI
 
 struct ContentView: View {
     var body: some View {
-        List(0..<10){_ in
-            HStack {
-                Rectangle().fill(Color("main")).frame(width: 5, height: 100, alignment: .leading)
-            
-                VStack{
-                    Spacer()
-                    HStack{
-                        Text("today")
-                    }
-                    Spacer()
-                    HStack {
-                        Text("Today")
-                    }
-                    Spacer()
-                }
-                VStack(alignment: .leading){
-                    Spacer()
-                    HStack{
-                        Text("programming").lineLimit(2)
-                    }
-                    Spacer()
-                    HStack{
-                        Text("Week 1 Description")
-                    }
-                    Spacer()
-                }
-                Spacer()
-                Text("w222")
-                
-                }.frame(height: 100).listRowInsets(EdgeInsets(top: -20, leading: -20, bottom: -20, trailing: -20))
-        }
+        LectureList()
     }
 }
